Update build configurations and add villager messaging functionality
This commit updates the version of the shadow plugin and the Java language version in the build configuration. Modifications have been made in the shadowJar configuration. The main feature in this commit is the introduction of a messaging system for villagers. This includes the necessary methods and fields in the existing villager classes and the addition of a new VillagerMessagesConfig class. Messages can now be added via config and each villager type can potentially return a random message when interacted with. A sound effect has also been added to correlate with the villager type. The WorthConfig class has been additionally refined to handle invalid material cases.